requirements.txt
paho-mqtt
fastapi
uvicorn
